backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.9.0.80
mediapipe==0.10.32
pillow>=11.0.0
requests==2.31.0
firebase-admin==6.4.0
python-dotenv==1.0.0
numpy==1.26.3
